Release #13 - Adjusting forms and implementing security
