+++ app.py
@@ -1,14 +1,19 @@
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
+from flask_wtf.csrf import CSRFProtect
+from flask_bcrypt import Bcrypt
 
 
 app = Flask(__name__)
 app.config.from_pyfile('config.py')
 
 db = SQLAlchemy(app)
+csrf = CSRFProtect(app)
+bcrypt = Bcrypt(app)
 
 
-from views import * # NoQA
+from views_game import * # NoQA
+from views_user import * # NoQA
 
 
 if __name__ == '__main__':
