+++ requirements.txt
@@ -1,7 +1,10 @@
+bcrypt==4.2.0
 blinker==1.8.2
 click==8.1.7
 Flask==3.0.3
+Flask-Bcrypt==1.0.1
 Flask-SQLAlchemy==3.1.1
+Flask-WTF==1.2.1
 greenlet==3.0.3
 importlib_metadata==8.2.0
 itsdangerous==2.2.0
@@ -13,4 +16,5 @@ python-dotenv==1.0.1
 SQLAlchemy==2.0.32
 typing_extensions==4.12.2
 Werkzeug==3.0.3
+WTForms==3.1.2
 zipp==3.20.0
