+++ templates/list.html
@@ -2,7 +2,7 @@
 {% block content %}
         <a class="btn btn-primary" href="{{ url_for('login') }}">Login</a>
         <a class="btn btn-danger" href="{{ url_for('logout') }}">Logout</a>
-        <a class="btn btn-primary" href="{{ url_for('new') }}">New Game</a>
+        <a class="btn btn-success" href="{{ url_for('new') }}">New Game</a>
         <table class="table table-striped table-responsive table-bordered">
             <thead class="thead-default">
                 <tr>
@@ -19,8 +19,8 @@
                     <td>{{ game.category }}</td>
                     <td>{{ game.console }}</td>
                     <td>
-                        <a href="{{ url_for('edit', id=game.id) }}">Edit</a>
-                        <a href="{{ url_for('delete', id=game.id) }}">Delete</a>
+                        <a class="btn btn-primary" href="{{ url_for('edit', id=game.id) }}">Edit</a>
+                        <a class="btn btn-danger" href="{{ url_for('delete', id=game.id) }}">Delete</a>
                     </td>
                 </tr>
                 {% endfor %}
