+++ templates/login.html
@@ -5,6 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>GameTeca</title>
     <link rel="stylesheet" href="{{ url_for('static', filename='bootstrap.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', filename='app.css') }}">
   </head>
   <body>
     <div class="container">
@@ -20,11 +21,20 @@
         <h1>Log in your account</h1>
         <form method="POST" action="{{ url_for('authenticate') }}">
             <input type="hidden" name="next" value="{{ next }}">
-            <p><label>Username:</label> <input class="form-control" type="text" name="username" required></p>
-            <p><label>Password:</label> <input class="form-control" type="password" name="password" required></p>
-            <p><button class="btn btn-primary" type="submit">Login</button></p>
+            {{ form.csrf_token() }}
+            <div class="form-group">
+                {{ form.username.label(class="form-label") }}
+                {{ form.username(class="form-control") }}
+            </div>
+            <div class="form-group">
+                {{ form.password.label(class="form-label") }}
+                {{ form.password(class="form-control") }}
+            </div>
+            <div class="form-group buttons">
+                {{ form.login(class="btn btn-primary") }}
+                <a class="btn btn-danger" href="{{ url_for('index') }}">Back</a>
+            </div>
         </form>
-        <a class="btn btn-danger" href="{{ url_for('index') }}">Back</a>
     </div>
 </body>
 </html>
