+++ templates/template.html
@@ -20,6 +20,9 @@
         {% endwith %}
       <div class="page-header">
           <h1>{{ title }}</h1>
+          {% if logged_msg %}
+          <h4>{{ logged_msg }}</h4>
+          {% endif %}
       </div>
       {% block content %}{% endblock %}
       <script type="text/javascript" src="{{ url_for('static', filename='jquery.js') }}"></script>
